requirements.txt
streamlit
google-genai
pandas
python-dotenv
supabase
